requirements.txt
asgiref==3.5.2
coverage==6.4.1
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
